# internal/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API request bodies and responses are usually JSON-encoded, except for binary data and some of the OAuth2 endpoints. Note that the `Content-Type` header must be set to `application/json` for this, as the request may otherwise fail with error 400.\nWhen clients have a valid access token, e.g. obtained through the `POST /api/v1/session` or `POST /api/v1/oauth/token` endpoint, they can use a standard Bearer Authorization header to authenticate their requests. Submitting the access token with a custom `X-Auth-Token` header is supported as well.",
        "title": "PhotoPrism API",
        "contact": {},
        "version": "v1"
    },
    "host": "demo.photoprism.app",
    "paths": {
...
        "/api/v1/photos": {
            "get": {
                "description": "Fore more information see:\n- https://docs.photoprism.app/developer-guide/api/search/#get-apiv1photos",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Photos"
                ],
                "summary": "finds pictures and returns them as JSON",
                "operationId": "SearchPhotos",
                "parameters": [
                    {
                        "maximum": 100000,
                        "minimum": 1,
                        "type": "integer",
                        "description": "maximum number of files",
                        "name": "count",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maximum": 100000,
                        "minimum": 0,
                        "type": "integer",
                        "description": "file offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "favorites",
                            "name",
                            "title",
                            "added",
                            "edited"
                        ],
                        "type": "string",
                        "description": "sort order",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "groups consecutive files that belong to the same photo",
                        "name": "merged",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "excludes private pictures",
                        "name": "public",
                        "in": "query"
                    },
                    {
                        "enum": [
                            0,
                            1,
                            2,
                            3,
                            4,
                            5,
                            6,
                            7
                        ],
                        "type": "integer",
                        "description": "minimum quality score (1-7)",
                        "name": "quality",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "search query",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "album uid",
                        "name": "s",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "photo path",
                        "name": "path",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "is type video",
                        "name": "video",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/search.Photo"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/photos/{uid}/files/{fileuid}/unstack": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Photos",
                    "Stacks"
                ],
                "summary": "removes a file from an existing photo stack",
                "operationId": "PhotoUnstack",
                "parameters": [
                    {
                        "type": "string",
                        "description": "photo uid",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "file uid",
                        "name": "fileuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Photo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/services/{id}/upload": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Services"
                ],
                "summary": "uploads files to the selected service account",
                "operationId": "UploadToService",
                "parameters": [
                    {
                        "type": "string",
                        "description": "service id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.File"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.Photo": {
            "type": "object",
            "properties": {
                "Albums": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Album"
                    }
                },
                "Altitude": {
                    "type": "integer"
                },
                "Camera": {
                    "$ref": "#/definitions/entity.Camera"
                },
                "CameraID": {
                    "type": "integer"
                },
                "CameraSerial": {
                    "type": "string"
                },
                "CameraSrc": {
                    "type": "string"
                },
                "Caption": {
                    "type": "string"
                },
                "CaptionSrc": {
                    "type": "string"
                },
                "Cell": {
                    "$ref": "#/definitions/entity.Cell"
                },
                "CellAccuracy": {
                    "type": "integer"
                },
                "CellID": {
                    "type": "string"
                },
                "Color": {
                    "type": "integer"
                },
                "Country": {
                    "type": "string"
                },
                "CreatedBy": {
                    "type": "string"
                },
                "Day": {
                    "type": "integer"
                },
                "Description": {
                    "type": "string"
                },
                "DescriptionSrc": {
                    "type": "string"
                },
                "Details": {
                    "$ref": "#/definitions/entity.Details"
                },
                "DocumentID": {
                    "type": "string"
                },
                "Duration": {
                    "$ref": "#/definitions/time.Duration"
                },
                "EstimatedAt": {
                    "type": "string"
                },
                "Exposure": {
                    "type": "string"
                },
                "FNumber": {
                    "type": "number"
                },
                "Faces": {
                    "type": "integer"
                },
                "Favorite": {
                    "type": "boolean"
                },
                "FocalLength": {
                    "type": "integer"
                },
                "Iso": {
                    "type": "integer"
                },
                "Lat": {
                    "type": "number"
                },
                "Lens": {
                    "$ref": "#/definitions/entity.Lens"
                },
                "LensID": {
                    "type": "integer"
                },
                "Lng": {
                    "type": "number"
                },
                "Month": {
                    "type": "integer"
                },
                "Name": {
                    "type": "string"
                },
                "OriginalName": {
                    "type": "string"
                },
                "Panorama": {
                    "type": "boolean"
                },
                "Path": {
                    "type": "string"
                },
                "Place": {
                    "$ref": "#/definitions/entity.Place"
                },
                "PlaceID": {
                    "type": "string"
                },
                "PlaceSrc": {
                    "type": "string"
                },
                "Private": {
                    "type": "boolean"
                },
                "PublishedAt": {
                    "type": "string"
                },
                "Quality": {
                    "type": "integer"
                },
                "Resolution": {
                    "type": "integer"
                },
                "Scan": {
                    "type": "boolean"
                },
                "Stack": {
                    "type": "integer"
                },
                "TakenAt": {
                    "type": "string"
                },
                "TakenAtLocal": {
                    "type": "string"
                },
                "TakenSrc": {
                    "type": "string"
                },
                "TimeZone": {
                    "type": "string"
                },
                "Title": {
                    "type": "string"
                },
                "TitleSrc": {
                    "type": "string"
                },
                "Type": {
                    "type": "string"
                },
                "TypeSrc": {
                    "type": "string"
                },
                "UID": {
                    "type": "string"
                },
                "Year": {
                    "type": "integer"
                },
                "checkedAt": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "editedAt": {
                    "type": "string"
                },
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.File"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "labels": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.PhotoLabel"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "i18n.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "search.Photo": {
            "type": "object",
            "properties": {
                "Altitude": {
                    "type": "integer"
                },
                "CameraID": {
                    "description": "Camera",
                    "type": "integer"
                },
                "CameraMake": {
                    "type": "string"
                },
                "CameraModel": {
                    "type": "string"
                },
                "CameraSerial": {
                    "type": "string"
                },
                "CameraSrc": {
                    "type": "string"
                },
                "CameraType": {
                    "type": "string"
                },
                "Caption": {
                    "type": "string"
                },
                "CellAccuracy": {
                    "type": "integer"
                },
                "CellID": {
                    "description": "Cell",
                    "type": "string"
                },
                "CheckedAt": {
                    "type": "string"
                },
                "Color": {
                    "type": "integer"
                },
                "Country": {
                    "type": "string"
                },
                "CreatedAt": {
                    "type": "string"
                },
                "Day": {
                    "type": "integer"
                },
                "DeletedAt": {
                    "type": "string"
                },
                "DocumentID": {
                    "type": "string"
                },
                "Duration": {
                    "$ref": "#/definitions/time.Duration"
                },
                "EditedAt": {
                    "type": "string"
                },
                "Exposure": {
                    "type": "string"
                },
                "FNumber": {
                    "type": "number"
                },
                "Faces": {
                    "type": "integer"
                },
                "Favorite": {
                    "type": "boolean"
                },
                "FileName": {
                    "type": "string"
                },
                "FileRoot": {
                    "type": "string"
                },
                "FileUID": {
                    "type": "string"
                },
                "Files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.File"
                    }
                },
                "FocalLength": {
                    "type": "integer"
                },
                "Hash": {
                    "type": "string"
                },
                "Height": {
                    "type": "integer"
                },
                "ID": {
                    "type": "string"
                },
                "InstanceID": {
                    "type": "string"
                },
                "Iso": {
                    "type": "integer"
                },
                "Lat": {
                    "type": "number"
                },
                "LensID": {
                    "description": "Lens",
                    "type": "integer"
                },
                "LensMake": {
                    "type": "string"
                },
                "LensModel": {
                    "type": "string"
                },
                "Lng": {
                    "type": "number"
                },
                "Merged": {
                    "type": "boolean"
                },
                "Month": {
                    "type": "integer"
                },
                "Name": {
                    "type": "string"
                },
                "OriginalName": {
                    "type": "string"
                },
                "Panorama": {
                    "type": "boolean"
                },
                "Path": {
                    "type": "string"
                },
                "PlaceCity": {
                    "type": "string"
                },
                "PlaceCountry": {
                    "type": "string"
                },
                "PlaceID": {
                    "type": "string"
                },
                "PlaceLabel": {
                    "type": "string"
                },
                "PlaceSrc": {
                    "type": "string"
                },
                "PlaceState": {
                    "type": "string"
                },
                "Portrait": {
                    "type": "boolean"
                },
                "Private": {
                    "type": "boolean"
                },
                "Quality": {
                    "type": "integer"
                },
                "Resolution": {
                    "type": "integer"
                },
                "Scan": {
                    "type": "boolean"
                },
                "Stack": {
                    "type": "integer"
                },
                "TakenAt": {
                    "type": "string"
                },
                "TakenAtLocal": {
                    "type": "string"
                },
                "TakenSrc": {
                    "type": "string"
                },
                "TimeZone": {
                    "type": "string"
                },
                "Title": {
                    "type": "string"
                },
                "Type": {
                    "type": "string"
                },
                "TypeSrc": {
                    "type": "string"
                },
                "UID": {
                    "type": "string"
                },
                "UpdatedAt": {
                    "type": "string"
                },
                "Width": {
                    "type": "integer"
                },
                "Year": {
                    "type": "integer"
                }
            }
        },
...
    },
...
}